Move roctx_trace_entry_t to a new header Definition of roctx_trace_entry_t has been moved to a new header. This new header will contain the definitions of all the <API>_trace_entry_t types used to flush the payloads of the events. The purpose of this modification is to have only one header that needs to be included by developers that want to implement new plugins with the interface proposal. Some fields types of the <API>_trace_entry_t structures are defined either in tool.cpp or in headers lying in /src/core (/src/core/trace_buffer.h). Since trace_buffer.h is removed once ROCTrace has been installed, the definition of these types has been moved to the new header and deleted from the ancient header to avoid redefinition issues.
